+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.2.2] - 2025-11-09
+- performance atan2 
+- parformance unstable sort (total 64%)
+
 ## [0.2.0] - 2025-11-09
 - Error enum for returning the exact cause.
 - Added test/benchmark data generators.